Вынес рендер диаграммы в утилиту
